feat(runloop) introduce plugins_loader module using Plugins in new DB
It includes a converter of the legacy schema format to the new one.

The conversion is not 100% compatible, due to difference in feature sets
of the old and new schema libraries, but the schemas can be made compatible
using the `new_type` entry, where a new-style type definition can be
specified for a field, including new-style validations.